VACMS-11519: Adds national vet center service descriptions to node:view
