+++ config/sync/core.entity_view_mode.taxonomy_term.vet_center_service.yml
@@ -1,0 +1,10 @@
+uuid: 591069f6-07ab-4494-904a-a64dc308e288
+langcode: en
+status: true
+dependencies:
+  module:
+    - taxonomy
+id: taxonomy_term.vet_center_service
+label: 'Vet Center service'
+targetEntityType: taxonomy_term
+cache: true
