+++ docroot/modules/custom/va_gov_vet_center/css/vet_center_service.css
@@ -1,0 +1,23 @@
+.centralized.tooltip-layout {
+  margin-bottom: 0;
+}
+
+.centralized.tooltip-layout .tooltip-toggle {
+  right: 2rem;
+}
+
+.centralized.tooltip-layout .field--name-field-vet-center-service-descrip {
+  color: #595959;
+  margin-top: 10px;
+  padding-top: 10px;
+}
+
+.field--name-field-health-services .field__items .field__item > *:not(:nth-child(1)) {
+  margin: 0 1rem;
+}
+
+.field--name-field-health-services .field__items .field__item ul {
+  padding-left: 1rem;
+}
+
+
