+++ docroot/modules/custom/va_gov_vet_center/va_gov_vet_center.libraries.yml
@@ -17,3 +17,8 @@ prevent_select_duplicates:
       css/prevent_select_duplicates.css: {}
   js:
     js/prevent_select_duplicates.js: {}
+display_service_descriptions:
+  css:
+    component:
+      css/vet_center_service.css: {}
+
