+++ docroot/modules/custom/va_gov_vet_center/va_gov_vet_center.services.yml
@@ -1,7 +1,7 @@
 services:
   va_gov_vet_center.entity_event_subscriber:
     class: Drupal\va_gov_vet_center\EventSubscriber\EntityEventSubscriber
-    arguments: ['@string_translation', '@va_gov_user.user_perms', '@va_gov_vet_center.required_services', '@entity_type.manager']
+    arguments: ['@string_translation', '@va_gov_user.user_perms', '@va_gov_vet_center.required_services', '@entity_type.manager', '@renderer']
     tags:
       - { name: event_subscriber }
   va_gov_vet_center.required_services:
